defconfig: Do not sync on suspend
Android builds up a great deal of cache and dirty pages for memory
instensive applications. If the user places the device into a sleep
state and the device is allowed to suspend, all of these pages will be
written to disk. Android applications may benefit from being allowed to
keep their pages in memory even after a suspend in terms of memory
performance and latency.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>